requirements.txt
fastapi[all]
passlib[bcrypt]
aiomysql
alembic
python-dotenv
cryptography
python-jose